requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
trl>=0.7.0
datasets>=2.12.0

# Training and optimization
bitsandbytes>=0.41.0
deepspeed>=0.10.0

# Attention kernels
triton>=2.0.0

# Evaluation
pytrec_eval>=0.5
beir>=2.0.0

# Utilities
pyyaml>=6.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.7.0
